修复best_model_checkpoint检查

+++ utils/callback.py
@@ -17,7 +17,7 @@ class SavePeftModelCallback(TrainerCallback):
             # 保存效果最好的模型
             best_checkpoint_folder = os.path.join(args.output_dir, f"{PREFIX_CHECKPOINT_DIR}-best")
             # 因为只保存最新5个检查点，所以要确保不是之前的检查点
-            if os.path.exists(state.best_model_checkpoint):
+            if state.best_model_checkpoint is not None and os.path.exists(state.best_model_checkpoint):
                 if os.path.exists(best_checkpoint_folder):
                     shutil.rmtree(best_checkpoint_folder)
                 shutil.copytree(state.best_model_checkpoint, best_checkpoint_folder)
